+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for `noreturn` routines, introduced in Delphi 13.
+- Support for `NameOf` intrinsic, introduced in Delphi 13.
+- Support for implicit `Self` in `Initialize` and `Finalize` operators, introduced in Delphi 13.
+- Support for `DCCARM64EC` toolchain, introduced in Delphi 13.1.
 - `NoreturnContract` analysis rule, which flags `noreturn` routines that return normally.
 - Support for `if` expressions (e.g. `X := if Foo then Bar else Baz;`), introduced in Delphi 13.